+++ archive/openai-functions-codeinterpreter/.env.example
@@ -1,0 +1,1 @@
+OPENAI_API_KEY=YOUR_OPENAI_API_KEY
