+++ archive/openai-functions-codeinterpreter/plugins/common/config.json
@@ -1,0 +1,3 @@
+{
+    "enabled": true
+}
