+++ archive/openai-functions-codeinterpreter/plugins/python/config.json
@@ -1,0 +1,3 @@
+{
+    "enabled": true
+}
