+++ archive/openai-functions-codeinterpreter/plugins/vue/config.json
@@ -1,0 +1,3 @@
+{
+    "enabled": true
+}
