--- openai-functions-codeinterpreter/.env.example
@@ -1,1 +1,0 @@
-OPENAI_API_KEY=YOUR_OPENAI_API_KEY
