--- openai-functions-codeinterpreter/plugins/common/config.json
@@ -1,3 +1,0 @@
-{
-    "enabled": true
-}
