--- openai-functions-codeinterpreter/plugins/python/config.json
@@ -1,3 +1,0 @@
-{
-    "enabled": true
-}
