--- openai-functions-codeinterpreter/plugins/vue/config.json
@@ -1,3 +1,0 @@
-{
-    "enabled": true
-}
